+++ backend/src/main/java/com/nuriwoolim/pagebackend/entity/TimeTable.java
@@ -1,6 +1,6 @@
 package com.nuriwoolim.pagebackend.entity;
 
-import com.nuriwoolim.pagebackend.global.BaseEntity;
+import com.nuriwoolim.pagebackend.core.BaseEntity;
 import com.nuriwoolim.pagebackend.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.ConstraintMode;
